requirements.txt
discord.py
transformers
torch
python-dotenv